complete admin validation

+++ Controllers/adminBrandController.js
@@ -1,25 +1,49 @@
 const adminBrand = require('../Model/adminBrand')
 
 const showBrandPage = (req,res)=>{
-  adminBrand.displayBrand(req.body).then((brands)=>{
-    res.render('admin/adminBrandPage',{admin:true,title:'BRAND CONTROL PAGE',brands})
-  })
+  if(req.session.admin)
+  {
+    adminBrand.displayBrand(req.body).then((brands)=>{
+      res.render('admin/adminBrandPage',{admin:true,title:'BRAND CONTROL PAGE',brands})
+    })
+  }
+  else
+  {
+    res.render('admin/adminLogin',{admin:false})
+  }
+
  
 }
 
 const addBrand = (req,res)=>{
-  adminBrand.insertbrand(req.body).then((response)=>{
-    res.redirect('/admin/adminBrandPage')
-  })
+  if(req.session.admin)
+  {
+    adminBrand.insertbrand(req.body).then((response)=>{
+      res.redirect('/admin/adminBrandPage')
+    })
+  }
+  else
+  {
+    res.render('admin/adminLogin',{admin:false})
+  }
+
 }
 
 
 
 const deleteBrand = (req,res)=>{
-  let brandId = req.query.id
-  adminBrand.deleteBrand(brandId).then((response)=>{
-    res.redirect('/admin/adminBrandPage')
-  })
+  if(req.session.admin)
+  {
+    let brandId = req.query.id
+    adminBrand.deleteBrand(brandId).then((response)=>{
+      res.redirect('/admin/adminBrandPage')
+    })
+  }
+  else
+  {
+    res.render('admin/adminLogin',{admin:false})
+  }
+
 }
 
 
@@ -30,5 +54,6 @@ const editBrand = (req,res)=>{
 module.exports = {
   showBrandPage,
   addBrand,
-  deleteBrand
+  deleteBrand,
+  editBrand
 }
