+++ Controllers/adminCategoryController.js
@@ -3,22 +3,45 @@ const adminCategory = require('../Model/adminCategory')
 
 
 const adminCategoryPage = (req,res)=>{
-  adminCategory.displayCategory().then((category)=>{
-    res.render('admin/adminCategoryPage',{admin:true,title:"CATEGORY CONTROL PAGE",category})
-  }) 
+  if(req.session.admin)
+  {
+      adminCategory.displayCategory().then((category)=>{
+      res.render('admin/adminCategoryPage',{admin:true,title:"CATEGORY CONTROL PAGE",category})
+    })
+  }
+  else
+  {
+    res.render('admin/adminLogin',{admin:false})
+  }
 }
 
 const addNewCategory = (req,res)=>{
-  adminCategory.insertcategory(req.body).then((response)=>{
-    res.redirect('/admin/adminCategoryPage')
-  })
+  if(req.session.admin)
+  {
+    adminCategory.insertcategory(req.body).then((response)=>{
+      res.redirect('/admin/adminCategoryPage')
+    })
+  }
+  else
+  {
+    res.render('admin/adminLogin',{admin:false})
+  }
+ 
 }
 
 const deleteCategory = (req,res)=>{
-  let categoryId = req.query.id
-  adminCategory.deleteCategory(categoryId).then((response)=>{
-    res.redirect('/admin/adminCategoryPage')
-  })
+  if(req.session.admin)
+  {
+    let categoryId = req.query.id
+    adminCategory.deleteCategory(categoryId).then((response)=>{
+      res.redirect('/admin/adminCategoryPage')
+    })
+  }
+  else
+  {
+    res.render('admin/adminLogin',{admin:false})
+  }
+
 }
